requirements.txt
fastapi==0.111.0
pydantic==2.7.0
qdrant-client==1.10.1
google-generativeai
python-dotenv